worked on auth and css

+++ src/dailytasks/views.py
@@ -8,11 +8,17 @@ from django.contrib.auth.decorators import login_required
 def task_view(request):
     if request.method == "POST":
         current_task = request.POST['addtask']
-        UserTask(task_name=current_task).save()
+        UserTask(user=request.user,task_name=current_task).save()
 
-    task_qs = UserTask.objects.all().order_by('completed','id')
+    task_qs = UserTask.objects.filter(user=request.user).order_by('completed','id')
+    number_of_tasks_completed = 0
+    for task in task_qs:
+        if task.completed == True:
+            number_of_tasks_completed += 1
+
     context = {
         'task_qs':task_qs,
+        'number_of_tasks_completed':number_of_tasks_completed,
     }
     return render(request,"tasks/task.html",context)
 
@@ -20,14 +26,13 @@ def task_view(request):
 def task_handle_view(request,task_id):
     current_task = UserTask.objects.get(id=task_id)
     
-    if request.method == "POST":
-        isDelete = request.POST.get('delete_task', False)
-        isComplete = request.POST.get('complete_task', False)
-        if isDelete:
-            current_task.delete()
-        elif isComplete:
-            current_task.completed = True
-            current_task.save()
-        
-    
+    if current_task.user == request.user:
+        if request.method == "POST":
+            isDelete = request.POST.get('delete_task', False)
+            isComplete = request.POST.get('complete_task', False)
+            if isDelete:
+                current_task.delete()
+            elif isComplete:
+                current_task.completed = True
+                current_task.save()
     return redirect("tasks")
